KP_style: landingpage home part finished
