+++ src/Navigation/Pages/LandingPage/staticData/data/containerdata.tsx
@@ -1,38 +1,51 @@
-export const containerData: any = {
-    container1: {
+interface DataTypes {
+    id: number;
+    title: string;
+    subtitle: string;
+    description: string;
+}
+
+export const containerData: DataTypes[] = [
+    {
+        id: 0,
         title: 'Welcome at front',
         subtitle: 'consectetur adipiscing elit, sed do eiusmod tempor ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
     },
-    container2: {
+    {
+        id: 1,
         title: '',
         subtitle: ' About us ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in.',
     },
-    container3: {
+    {
+        id: 2,
         title: '',
         subtitle: ' services ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in.',
     },
-    container4: {
+    {
+        id: 3,
         title: '',
         subtitle: ' App ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in.',
     },
-    container5: {
+    {
+        id: 4,
         title: '',
         subtitle: ' Contact ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in.',
     },
-    container6: {
+    {
+        id: 5,
         title: '',
         subtitle: ' More info ',
         description:
             'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in.',
     },
-};
+];
